Move DraftailRichTextArea definition to Draftail component and register it in entrypoints/admin/draftail.js
The definition of DraftailRichTextArea needs to be in index.js because we want to export it as part of the top-level `draftail` API (rather than having to import a submodule of components/Draftail when registering it with telepath), but it also depends on initEditor.
